updated testcases and resolve comments
