--- network_schema/network_schema.go
@@ -1,9 +1,0 @@
-package network_schema
-
-import (
-	gen "github.com/ibm-hyper-protect/contract-go/common/general"
-)
-
-func HpcrVerifyNetworkConfig(network_config string) error {
-	return gen.VerifyNetworkSchema(network_config)
-}
