+++ samples/network/network_config.yaml
@@ -1,0 +1,11 @@
+network:
+  version: 2
+  ethernets:
+    enc1:
+      dhcp4: false
+      addresses:
+        - 192.168.122.110/24 
+      gateway4: 192.168.122.1
+      nameservers:
+        addresses:
+          - 8.8.8.8
